Add fp32 support to all kernels (#66)
This PR adds support for `float32` to all kernels. Most kernels work
fine, however some fail to validate, and some others don't actually do
calculations in float64. See the spreadsheet for more details.

This PR also changes the default precision to float32 in all kernels
with floating-point precision.

I would propose to merge this pretty quickly and try to fix up in future
PRs

+++ npbench/benchmarks/azimint_hist/azimint_hist.py
@@ -2,7 +2,7 @@
 import numpy as np
 
 
-def initialize(N, datatype=np.float64):
+def initialize(N, datatype=np.float32):
     from numpy.random import default_rng
     rng = default_rng(42)
     data, radius = rng.random((N, ), dtype=datatype), rng.random((N, ), dtype=datatype)
